# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The JSON interface to a HyperUplink board, exposing the same\ncategories, forums, topics and administrative surfaces the web\ninterface serves, with every response shaped by the permissions\nof the user the API key belongs to.\n\nAuthentication happens with an API key that a signed-in user\nmints under Account -\u003e API on the board itself, and the secret\nit hands out once is sent either as a bearer token in the\nAuthorization header or verbatim in the X-API-Key header, so a\nrequest that carries neither is answered with 401 before it\never reaches a controller.",
        "title": "HyperUplink API",
        "contact": {
            "name": "HyperUplink",
            "url": "https://hyperup.link"
        },
        "license": {
            "name": "SEGV",
            "url": "https://xn--gckvb8fzb.com/segv/"
        },
        "version": "{{.Version}}"
    },
    "basePath": "/",
    "paths": {
...
        "/account/settings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Show the account settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "timezones": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "user": {
                                    "$ref": "#/definitions/user.Detail"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Update the account settings",
                "parameters": [
                    {
                        "description": "The settings to store",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/settings.UpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
        "/account/settings/view/{view}": {
            "post": {
                "description": "Flips whether the named section of the interface starts out open or\nclosed for the caller and answers with the state it now has.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Toggle a collapsible view",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The name of the view to toggle",
                        "name": "view",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "enabled": {
                                    "type": "boolean"
                                },
                                "view": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
        "/account/twofactor/disable": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Disable two-factor authentication",
                "parameters": [
                    {
                        "description": "The current password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/twofactor.TwofactorDisableInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
        "/admin/auth": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Show the authentication settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "auth": {
                                    "$ref": "#/definitions/setting.Auth"
                                },
                                "auth_providers": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update the authentication settings",
                "parameters": [
                    {
                        "description": "The authentication settings to store",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.UpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
        "/admin/board/attachments": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Show the attachment settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/attachments.View"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update the attachment settings",
                "parameters": [
                    {
                        "description": "The attachment settings to store",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/attachments.UpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "attachments.UpdateInput": {
            "type": "object",
            "required": [
                "max_size"
            ],
            "properties": {
                "enable_attachments": {
                    "type": "boolean"
                },
                "inline_image_display": {
                    "type": "boolean"
                },
                "max_size": {
                    "type": "integer",
                    "minimum": 1
                },
                "on_upload_hook": {
                    "type": "string",
                    "maxLength": 1024
                },
                "storage_path": {
                    "type": "string",
                    "maxLength": 255
                },
                "storage_provider_id": {
                    "type": "string",
                    "maxLength": 64
                },
                "upload_formats": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "attachments.View": {
            "type": "object",
            "properties": {
                "attachments": {
                    "$ref": "#/definitions/setting.Attachments"
                },
                "storage_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "storage_is_public": {
                    "type": "boolean"
                },
                "upload_format_options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "auth.UpdateInput": {
            "type": "object",
            "properties": {
                "address_type": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2
                    ]
                }
            }
        },
...
        "pgtype.InfinityModifier": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                0,
                -1
            ],
            "x-enum-varnames": [
                "Infinity",
                "Finite",
                "NegativeInfinity"
            ]
        },
...
        "pgtype.Timestamp": {
            "type": "object",
            "properties": {
                "infinityModifier": {
                    "$ref": "#/definitions/pgtype.InfinityModifier"
                },
                "time": {
                    "description": "Time zone will be ignored when encoding to PostgreSQL.",
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
        "request.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "forbidden"
                }
            }
        },
        "request.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "request.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "setting.AddressType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "EmailOnly",
                "JID",
                "EmailAndJID"
            ]
        },
        "setting.Attachments": {
            "type": "object",
            "properties": {
                "enable_attachments": {
                    "type": "boolean"
                },
                "inline_image_display": {
                    "description": "a.k.a. \"Image Board\" mode",
                    "type": "boolean"
                },
                "max_size": {
                    "description": "in byte",
                    "type": "integer"
                },
                "on_upload_hook": {
                    "type": "string"
                },
                "storage_path": {
                    "type": "string"
                },
                "storage_provider_id": {
                    "type": "string"
                },
                "upload_formats": {
                    "description": "Options: image/gif, image/jpeg, image/png, image/webp, application/json, application/pdf, application/zip, audio/mpeg, audio/vorbis, video/mp4, text/plain, text/csv, text/html",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "setting.Auth": {
            "type": "object",
            "properties": {
                "allowed_address_type": {
                    "description": "0: EmailOnly, 1: JID, 2: Email and JID",
                    "allOf": [
                        {
                            "$ref": "#/definitions/setting.AddressType"
                        }
                    ]
                }
            }
        },
...
        "settings.UpdateInput": {
            "type": "object",
            "required": [
                "timezone"
            ],
            "properties": {
                "timezone": {
                    "type": "string"
                }
            }
        },
...
        "twofactor.TwofactorDisableInput": {
            "type": "object",
            "required": [
                "current_password"
            ],
            "properties": {
                "current_password": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8
                }
            }
        },
...
        "user.Detail": {
            "type": "object",
            "properties": {
                "banned_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "confirmed_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "created_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "deleted_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "email": {
                    "type": "string"
                },
                "email_is_jid": {
                    "type": "boolean"
                },
                "email_unconfirmed": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "member_of": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "otp_enabled": {
                    "type": "boolean"
                },
                "profile_picture": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/user.Role"
                },
                "sign_in_last_at": {
                    "$ref": "#/definitions/pgtype.Timestamp"
                },
                "signature_html": {
                    "type": "string"
                },
                "signature_text": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "user.Role": {
            "type": "string",
            "enum": [
                "guest",
                "user",
                "admin"
            ],
            "x-enum-varnames": [
                "GuestRole",
                "UserRole",
                "AdminRole"
            ]
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "The API key secret on its own.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "The API key secret prefixed with \"Bearer \".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}